[web] - add ability to create simple task
