+++ web/.env.development
@@ -1,7 +1,7 @@
 PUBLIC_URL=/
 REACT_APP_GOSH_NETWORK="http://localhost"
-REACT_APP_GOSH_ROOTADDR="0:24476515d149782567696de0d590998b8f020e07c858926d21e6c8713588d16e"
-REACT_APP_GOSH='{"6.2.0": "0:a8990d70f9997591228a07ee892bf569134458012652fa862a38030ae00c0b14"}'
+REACT_APP_GOSH_ROOTADDR="0:93609d15d0dee074a16a033c1cea28f18cf589701b34defe92282849564a5d09"
+REACT_APP_GOSH='{"6.2.0": "0:3d52b798f7571e8ed66a3e04ae9210fc3399501ed3292f6bbc1a9eac8cfe4089"}'
 REACT_APP_IPFS="https://ipfs.network.gosh.sh"
 REACT_APP_GOSHAI_PROFILE="0:3b3484aa0585b43397e864d74cf20df53e20411201573344910a155aff61e507"
 REACT_APP_GOSHAI_NAME="gosh-ai"
