+++ web/src/v6.2.0/components/Modal/TaskCreate/index.ts
@@ -1,0 +1,1 @@
+export * from './TaskCreate'
