+++ web/src/v6.2.0/components/Modal/index.ts
@@ -1,10 +1,11 @@
-export * from './RepositoryCreate/RepositoryCreate'
-export * from './DaoUpgradeComplete/DaoUpgradeComplete'
+export * from './DaoMemberOf/DaoMemberOf'
 export * from './DaoTokenMint/DaoTokenMint'
 export * from './DaoTokenSend/DaoTokenSend'
+export * from './DaoUpgradeComplete/DaoUpgradeComplete'
 export * from './MemberTokenSend/MemberTokenSend'
-export * from './RequestDaoMembership/RequestDaoMembership'
-export * from './DaoMemberOf/DaoMemberOf'
-export * from './PinCode'
 export * from './MilestoneCreate/MilestoneCreate'
 export * from './MilestoneCreate/MilestoneTaskCreate'
+export * from './PinCode'
+export * from './RepositoryCreate/RepositoryCreate'
+export * from './RequestDaoMembership/RequestDaoMembership'
+export * from './TaskCreate'
