--- web/src/v6.2.0/pages/TaskCreate/index.ts
@@ -1,2 +1,0 @@
-export * from './TaskCreate'
-export { default } from './TaskCreate'
